Favoring Huborg over local development
In working on the Samvera class that I just removed, I found that the
methods/functions I was using for Samvera were generally applicable at
my local institution.

I spent a bit of time extracting and expanding that behavior, and
created the [huborg][1] gem. Now, each institution can benefit from the
Github organization methods that we the Samvera community leverage. The
Huborg gem provides a place to share those approaches.

[1]:https://github.com/samvera-labs/huborg