Updating documentation & behavior of script
Prior to this commit, I had assumed that the script would work when ran
from anywhere. But alas, Bundler applies some smarts, and doesn't let a
CLI task change context. So, for the time being, you'll need to copy the
script into the bundled directory.

Also, as noted in the comments, I'm now checking the spec based on the
version that bundler has bundled up.

Prior to this commit, if bundler said to use v1.0 of a gem and I had
v1.0 and v1.1 on my machine, the gem specification would pick up
v1.1. This commit ensures that we read the specification of the bundled
version.